Refactor lib service
