--- lib/common/service_test.go
@@ -1,72 +1,0 @@
-/*
- *  Copyright 2020 KardiaChain
- *  This file is part of the go-kardia library.
- *
- *  The go-kardia library is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  The go-kardia library is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with the go-kardia library. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package common
-
-import (
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/require"
-)
-
-type testService struct {
-	BaseService
-}
-
-func (testService) OnReset() error {
-	return nil
-}
-
-func TestBaseServiceWait(t *testing.T) {
-	ts := &testService{}
-	ts.BaseService = *NewBaseService(nil, "TestService", ts)
-	ts.Start()
-
-	waitFinished := make(chan struct{})
-	go func() {
-		ts.Wait()
-		waitFinished <- struct{}{}
-	}()
-
-	go ts.Stop()
-
-	select {
-	case <-waitFinished:
-		// all good
-	case <-time.After(100 * time.Millisecond):
-		t.Fatal("expected Wait() to finish within 100 ms.")
-	}
-}
-
-func TestBaseServiceReset(t *testing.T) {
-	ts := &testService{}
-	ts.BaseService = *NewBaseService(nil, "TestService", ts)
-	ts.Start()
-
-	err := ts.Reset()
-	require.Error(t, err, "expected cant reset service error")
-
-	ts.Stop()
-
-	err = ts.Reset()
-	require.NoError(t, err)
-
-	err = ts.Start()
-	require.NoError(t, err)
-}
